refactor: extract IPCMixin from NexusFS (pipe/stream internal methods)
Move _pipe_read, _pipe_write, _pipe_destroy, _stream_read,
_stream_write, _stream_destroy to nexus_fs_ipc.py as IPCMixin.

~110 lines extracted. Same mixin pattern as Watch/Lock.
